server/requirements.txt
black
python-dotenv
requests
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pyjwt
daphne
channels
channels-redis
redis
psycopg2-binary